requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2==2.9.5
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
